chore: unified annotation keywords (#233)
* chore: unified annotation keywords

* chore: moved to a shared place
