+++ properties_pane/view_level/viewLevelConfig.json
@@ -703,7 +703,7 @@
 				"structure": [
 					{
 						"propertyName": "Name",
-						"propertyKeyword": "viewAnnotationName",
+						"propertyKeyword": "annotationName",
 						"propertyTooltip": "An identifier that can have up to 1024 characters. If the annotation name is a reserved word it must be provided in double quotes. When a double quoted identifier is used, the identifier can also contain whitespace characters",
 						"propertyType": "text",
 						"validation": {
@@ -712,7 +712,7 @@
 					},
 					{
 						"propertyName": "Value",
-						"propertyKeyword": "viewAnnotationValue",
+						"propertyKeyword": "annotationValue",
 						"propertyTooltip": "The value of the annotation is optionsal.  Value can have a maximum of 4000 characters.",
 						"propertyType": "details",
 						"template": "textarea",
